chore(feed): break out big helper functions in own files
